Added local storage and change Float to BigDecimal

+++ base-ui-kit/src/main/java/com/vinapp/base_ui_kit/qouteitem/QuoteItem.kt
@@ -244,8 +244,8 @@ private fun QuoteItemPreview() {
                 percentageChange = 4.56F,
                 exchangeName = "MCX",
                 securityName = "Asdfghjkl",
-                lastTradePrice = 123.45678F,
-                changeInPrice = 0.987F
+                lastTradePrice = 123.45678F.toBigDecimal(),
+                changeInPrice = 0.987F.toBigDecimal()
             )
         )
     )
