+++ domain/src/main/java/com/vinapp/domain/entity/UpdatableQuote.kt
@@ -1,10 +1,12 @@
 package com.vinapp.domain.entity
 
+import java.math.BigDecimal
+
 data class UpdatableQuote(
     val ticker: String,
     val percentageChange: Float?,
     val exchangeName: String?,
     val securityName: String?,
-    val lastTradePrice: Float?,
-    val changeInPrice: Float?
+    val lastTradePrice: BigDecimal?,
+    val changeInPrice: BigDecimal?
 )
